backend/requirements.txt
django~=4.2
djangorestframework~=3.14
psycopg2-binary~=2.9
dj-database-url~=2.1
python-dotenv~=1.0
drf-spectacular~=0.27
waitress~=2.1
